requirements.txt
gunicorn
django
djangorestframework
pylibmc
# Add other dependencies your application needs
